requirements.txt
# ---- LLM ----
openai>=1.40,<2.0

# ---- Embedding / vector retrieval ----
# numpy must be <2, otherwise faiss-cpu 1.7.x and sentence-transformers 2.x will crash
numpy>=1.24,<2.0
# faiss-cpu only officially supports numpy 2.x starting from 1.8; to stay paired with numpy<2, pin to 1.7.4
faiss-cpu>=1.7.4,<1.9
# sentence-transformers 2.x requires huggingface_hub<0.26 (0.26 removed cached_download)
sentence-transformers>=2.7,<3.0
huggingface_hub==0.25.2
# Newer transformers versions changed the tokenizer interface; pin a range verified with st 2.7
transformers>=4.41,<4.46
# torch CPU build; st will pull it automatically but not pinning leads to version drift
torch>=2.1,<2.5

# ---- UI (versions must be pinned, this is the golden compatible combo) ----
gradio==4.44.1
gradio_client==1.3.0
# gradio 4.44 depends on pydantic>=2; 2.10+ has sporadic schema issues, so pin <2.10
pydantic>=2.7,<2.10

# ---- Utilities ----
python-dotenv>=1.0

# ---- Testing ----
pytest>=7.0
